requirements.txt
# Microsoft Sentinel MCP Server - Python Dependencies
# Generated: 2025-11-19
# Python Version: >=3.10

# Core MCP Framework
fastmcp>=0.1.0  # FastMCP - Simple Pythonic MCP server implementation

# Azure SDK
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-securityinsight>=2.0.0b2
azure-monitor-query>=1.2.0

# Microsoft Authentication Library
msal>=1.26.0

# HTTP & API
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# PowerShell Integration
# Note: PowerShell Core must be installed separately
# See: https://learn.microsoft.com/en-us/powershell/scripting/install/installing-powershell

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Logging & Monitoring
structlog>=23.2.0

# CLI
click>=8.1.0
rich>=13.7.0  # For beautiful CLI output

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Development
ipython>=8.18.0
ipdb>=0.13.13

# Security
bandit>=1.7.5  # Security linting
safety>=2.3.5  # Dependency vulnerability scanning
